requirements.txt
absl-py==2.3.0
astunparse==1.6.3
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cuda-bindings==12.8.0
cupti-python==12.8.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
joblib==1.5.1
kaggle==1.7.4.5
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
networkx==3.4.2
numpy==2.1.3
nvidia-cuda-cupti-cu12==12.9.19
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.4
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.3
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.26
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
